+++ cryptolibs/clefia.c
@@ -13,7 +13,7 @@ void setKey(unsigned char* key, int keybits)
 	_nrounds = ClefiaKeySet(_rk, key, keybits);
 }
 
-void xor(unsigned int* a, const unsigned int* b) {
+void xor(unsigned long* a, const unsigned long* b) {
 	int i = 5;
 	while (--i > 0) {
 		*a ^= *b;
@@ -24,7 +24,7 @@ void xor(unsigned int* a, const unsigned int* b) {
 
 void encrypt(const unsigned char plaintext[16], unsigned char ciphertext[16])
 {
-	xor( (unsigned int*) _ive, (unsigned int*) plaintext);
+	xor( (unsigned long*) _ive, (unsigned long*) plaintext);
 	ClefiaEncrypt(ciphertext, _ive, _rk, _nrounds);
 	memcpy(_ive, ciphertext, 16);
 }
@@ -32,6 +32,6 @@ void encrypt(const unsigned char plaintext[16], unsigned char ciphertext[16])
 void decrypt(const unsigned char ciphertext[16], unsigned char plaintext[16])
 {
 	ClefiaDecrypt(plaintext, ciphertext, _rk, _nrounds);
-	xor( (unsigned int*) plaintext, (unsigned int*) _ivd);
+	xor( (unsigned long*) plaintext, (unsigned long*) _ivd);
 	memcpy(_ivd, ciphertext, 16);
 }
